+++ tests/unit/lvelementstest/jsfunctiontest.cpp
@@ -55,7 +55,7 @@ class FunctionStub : public Element{
 
     static int variableParametersType(const Function::CallInfo& p){
         forwardCall()(p.at(0).toInt32(p.engine()), p.at(1).toNumber(p.engine()), p.at(2).toStdString(p.engine()));
-        return p.length();
+        return (int)p.length();
     }
 
     static void variableParametersVoidType(const Function::CallInfo& p){
